+++ tests/fixtures/sample_solutions/java/TwoSum.java
@@ -1,0 +1,23 @@
+/**
+ * 1. Two Sum
+ *
+ * Given an array of integers nums and an integer target, return indices
+ * of the two numbers such that they add up to target.
+ */
+
+import java.util.HashMap;
+import java.util.Map;
+
+class Solution {
+    public int[] twoSum(int[] nums, int target) {
+        Map<Integer, Integer> seen = new HashMap<>();
+        for (int i = 0; i < nums.length; i++) {
+            int complement = target - nums[i];
+            if (seen.containsKey(complement)) {
+                return new int[] { seen.get(complement), i };
+            }
+            seen.put(nums[i], i);
+        }
+        return new int[] {};
+    }
+}